sdm660-common: enable navbar with overlay instead
As for now qemu hardware mainkeys property is set for that purpose, but for a reason navbar settings are
shown under system/buttons (can see on twelve-plus), hence not avaliable on regular twelve. Let`s use overlay to enable
navbar, so it`s settings should be shown under system/gestures menu.